# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Austin Rifle Club API",
    "description": "API for Austin Rifle Club member portal, range status, events, and administration.",
    "version": "1.0.0",
    "contact": {
      "email": "info@austinrifleclub.org"
    }
  },
  "servers": [
    {
      "url": "https://austinrifleclub.org/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787/api",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/range-status/{id}": {
      "get": {
        "tags": ["Range Status"],
        "summary": "Get status of a specific range",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Range status" },
          "404": { "description": "Range not found" }
        }
      },
      "patch": {
        "tags": ["Range Status"],
        "summary": "Update range status (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["open", "closed", "event", "maintenance"] },
                  "statusNote": { "type": "string" },
                  "expiresAt": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated range status" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Range not found" }
        }
      }
    },
    "/range-status/{id}/close": {
      "post": {
        "tags": ["Range Status"],
        "summary": "Quick close a range (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reason"],
                "properties": {
                  "reason": { "type": "string" },
                  "duration": { "type": "integer", "description": "Duration in minutes" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Range closed" }
        }
      }
    },
    "/range-status/{id}/open": {
      "post": {
        "tags": ["Range Status"],
        "summary": "Quick open a range (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Range opened" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}